requirements.txt
requests
pydocx